Another renderers update.
* NEW: Script shaders now can use all 6 shader elements like in engine
blenders and explicitly declare them without using named indexes like
"normal_hq" or "l_special". Use "function element_X" where X is element
index [0-5].

* FIX: Fixed color bleeding of front objects into distortion (particles
are still have wrong sorting though).
* FIX: Fixed motion blur.

* ADD: Exported inverted view and inverted projection matrices to
shaders (m_invV, m_invP).
* ADD: Exported wind direction and intensity params to shaders
(wind_params).
* ADD: Implemented R3\R4 CRenderTarget::disable_aniso (+5 fps).

* EDIT: Script shaders now use ".lua" extension.
* EDIT: R2 doesn't use legacy shader compiler anymore. It allow us to
use all features of SM3.0. All R2 shaders changed to use full precision
instead of half.
* EDIT: FXAA moved after combine_2 to work better (including final image
with bloom, distortion, etc).
* EDIT: Various minor optimizations & improvements, code cleanup.
* EDIT: Moved some blenders from engine to script shaders.

+++ game/gamedata/shaders/r2/ps_ssss_mrmnwar_mask_blur.hlsl
@@ -3,9 +3,9 @@
 
 #define MASK_BLUR_SAMPLES int(6)
 
-half4 main (p_screen I) : COLOR
+float4 main (p_screen I) : COLOR
 {
-	half4 outColor = half4(0,0,0,0);
+	float4 outColor = float4(0,0,0,0);
 	
 	for (int i = 1; i < MASK_BLUR_SAMPLES; ++i)
 	{
